fix cassandra clear/init no keyspace exception
